desigpt/backend/requirements.txt
fastapi
uvicorn
python-dotenv
agno-agi
openai
